+++ examples/complete/outputs.tf
@@ -1,14 +1,14 @@
-output "this_ecs_instance_public_ip" {
-  description = "The ecs instance public ip."
-  value       = module.market-jenkins.this_ecs_instance_public_ip
+output "this_url_of_market_jenkins_with_ecs" {
+  description = "The url of jenkins."
+  value       = module.market_jenkins_with_ecs.this_jenkins_url
 }
 
-output "this_slb_public_address" {
-  description = "The slb public address."
-  value       = module.market-jenkins.this_slb_public_address
+output "this_url_of_market_jenkins_with_slb" {
+  description = "The url of jenkins with creating slb."
+  value       = module.market_jenkins_with_slb.this_jenkins_url
 }
 
-output "this_wordpress_url" {
-  description = "The url of wordpress."
-  value       = module.market-jenkins.this_wordpress_url
+output "this_url_of_market_jenkins_with_bind_dns" {
+  description = "The url of jenkins with binding dns."
+  value       = module.market_jenkins_with_bind_dns.this_jenkins_url
 }