+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 IMPROVEMENTS：
 
+- add more outputs and improve readme and examples [GH-2]( https://github.com/terraform-alicloud-modules/terraform-alicloud-market-jenkins/pull/2)
 - **NEW:** `market-jenkins` [GH-1]( https://github.com/terraform-alicloud-modules/terraform-alicloud-market-jenkins/pull/1)